requirements.txt
 streamlit
PyMuPDF
faiss-cpu
sentence-transformers
transformers
torch
accelerate
